scripts/requirements.txt
pyyaml==6.0.2
click
requests
paramiko
deepdiff